@@ -1,9 +1,8 @@
 import { sveltekit } from '@sveltejs/kit/vite'
 import mdsvexamples from 'mdsvexamples/vite'
-import type { UserConfig } from 'vite'
-import type { UserConfig as VitestConfig } from 'vitest'
+import { defineConfig } from 'vitest/config'
 
-export default {
+export default defineConfig(({ mode }) => ({
   plugins: [sveltekit(), mdsvexamples],
 
   test: {
@@ -11,6 +10,7 @@ export default {
     css: true,
     coverage: {
       reporter: [`text`, `json-summary`],
+      include: [`src/lib/Toc.svelte`],
     },
   },
 
@@ -22,4 +22,8 @@ export default {
   preview: {
     port: 3000,
   },
-} satisfies UserConfig & { test: VitestConfig }
+
+  resolve: {
+    conditions: mode === `test` ? [`browser`] : [],
+  },
+}))
